+++ src/main/java/com/altoros/blockchain/reinsurance/Financial.java
@@ -1,24 +1,27 @@
 package com.altoros.blockchain.reinsurance;
 
+import java.math.BigDecimal;
+
 /**
  * @author Nikita Gorbachevski
  */
 public class Financial {
-    private final long priority;
-    private final long limit;
+
+    private final BigDecimal priority;
+    private final BigDecimal limit;
     private final String priorityType;
 
-    public Financial(long priority, long limit, String priorityType) {
+    public Financial(BigDecimal priority, BigDecimal limit, String priorityType) {
         this.priority = priority;
         this.limit = limit;
         this.priorityType = priorityType;
     }
 
-    public long getPriority() {
+    public BigDecimal getPriority() {
         return priority;
     }
 
-    public long getLimit() {
+    public BigDecimal getLimit() {
         return limit;
     }
